add code and docker config for metakg-opposite-triples analysis
Running the metakg-opposite-triples.Dockerfile container extracts subject-predicate-object triple pairs from the MetaKG that share the same subject and object but use predicates that are opposite in nature. The opposite predicates are defined in the predicates/predicates.txt file (which is not in the branch but is on the main branch currently). The output is placed in classes/results/metakg-opposite-triples/metakg-opposite-triples.tsv.